crud/requirements.txt
flask
flask-login
python-dotenv
psycopg2-binary